remove trailing slashes from registries
This is a follow-up change for https://github.com/containers/image/pull/428.

It is a common pitfall to add a trailing slash when specifying a
registry (e.g., "docker.io/library/" version "docker.io/library").
Remove trailing slashes to normalize the representation of registries.

To facilitate testing, don't overwrite the BINARY and MIGRATOR variable
when being set in test.bats.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>